Updated to version 7.11.0. Currently without any ApiDefinition changes - this is meant as a quick fix for the iOS 10 error on submit:
This app attempts to access privacy-sensitive data without a usage description. The app's Info.plist must contain an NSCalendarsUsageDescription key with a string value explaining to the user how the app uses this data.
This app attempts to access privacy-sensitive data without a usage description. The app's Info.plist must contain an NSMotionUsageDescription key with a string value explaining to the user how the app uses this data.

See:

https://forums.developer.apple.com/thread/62229